k8s 环境初始化脚本

+++ scripts/os_setting.sh
@@ -2,27 +2,26 @@
 function os::setting::stop-firewall() {
     systemctl stop nftable
     systemctl disable nftable
-    iptables -F && iptables -X && iptables -F -t nat && iptables -X -t nat
-    iptables -P FORWARD ACCEPT
+    sudo /usr/sbin/iptables -F && sudo /usr/sbin/iptables -X && sudo /usr/sbin/iptables -F -t nat && sudo /usr/sbin/iptables -X -t nat
+    sudo /usr/sbin/iptables -P FORWARD ACCEPT
 }
 
 # 关闭 swap 分区，否则 kubelet 会启动失败 （设置 kubelet 启动参数 --fail-swap-on false 关闭 swap 检查）
 function os::setting::stop-swap() {
-    swapoff -a
-    sed -i '/ swap / s/^\(.*\)$/#\1/g' /etc/fstab
+    sudo swapoff -a
+    sudo sed -i '/ swap / s/^\(.*\)$/#\1/g' /etc/fstab
 }
 
 # 关闭 SELinux，负责 kubelet 挂在目录可能显示权限不足
 function os::setting::stop-selinux() {
-    setenforce 0
-    sed -i 's/^SELINUX=.*$/SELINUX=disabled/' /etc/selinux/config
+    sudo setenforce 0
+    sudo sed -i 's/^SELINUX=.*$/SELINUX=disabled/' /etc/selinux/config
 }
 
-# 优化内核参数，关闭 tcp_tw_recycle，负责和 NAT 冲突，导致服务不通
+# 优化内核参数，关闭 net.ipv4.tcp_tw_recycle=0，负责和 NAT 冲突，导致服务不通
 function os::setting::sysctl() {
-  cat kubernetes.conf <<EOF
+  cat > kubernetes.conf <<EOF
 net.ipv4.ip_forward=1
-net.ipv4.tcp_tw_recycle=0
 net.ipv4.neigh.default.gc_thresh1=1024
 net.ipv4.neigh.default.gc_thresh2=2048
 net.ipv4.neigh.default.gc_thresh3=4096
@@ -36,8 +35,8 @@ fs.nr_open=52706963
 net.ipv6.conf.all.disable_ipv6=1
 EOF
 
-cp kubernetes.conf /etc/sysctl.d/kubernetes.conf
-sysctl -p /etc/sysctl.d/kubernetes.conf
+sudo cp kubernetes.conf /etc/sysctl.d/kubernetes.conf
+sudo /usr/sbin/sysctl -p /etc/sysctl.d/kubernetes.conf
 }
 
 # 设置时间
@@ -50,9 +49,16 @@ function os::setting::time() {
   # 当前 UTC 时间写入硬件
   timedatectl set-local-rtc 0
   # 重启依赖系统时间的服务
-  systemctl restart rsyslog
+   for log_service in rsyslog syslog-ng systemd-journald; do
+      if systemctl list-unit-files | grep -q $log_service; then
+        systemctl try-restart $log_service
+        break
+      fi
+  done
   systemctl restart chrony
   # 关闭无关服务
-  systemctl stop postfix
-  systemctl disable postfix
+  if systemctl list-unit-files | grep -q postfix; then
+      systemctl stop postfix
+      systemctl disable postfix
+  fi
 }
